# Improve performance in program bottom menu evaluation