--- a/app/src/main/java/org/dhis2/usescases/searchTrackEntity/SearchPageConfigurator.kt
+++ b/app/src/main/java/org/dhis2/usescases/searchTrackEntity/SearchPageConfigurator.kt
@@ -1,6 +1,7 @@
 package org.dhis2.usescases.searchTrackEntity
 
 import org.dhis2.utils.customviews.navigationbar.NavigationPageConfigurator
+import org.hisp.dhis.android.core.common.FeatureType
 
 class SearchPageConfigurator(
     val searchRepository: SearchRepository,
@@ -10,7 +11,7 @@ class SearchPageConfigurator(
     private var canDisplayAnalytics: Boolean = false
 
     fun initVariables(): SearchPageConfigurator {
-        canDisplayMap = searchRepository.programHasCoordinates()
+        canDisplayMap = programHasCoordinates()
         canDisplayAnalytics = searchRepository.programHasAnalytics()
         return this
     }
@@ -29,5 +30,42 @@ class SearchPageConfigurator(
 
     override fun displayAnalytics(): Boolean {
         return canDisplayAnalytics
+    }
+
+    internal fun programHasCoordinates(): Boolean {
+        val program = searchRepository.currentProgram()?.let {programId ->
+            searchRepository.getProgram(programId)
+        } ?: return false
+
+        val programHasCoordinates = program.featureType() != null && program.featureType() != FeatureType.NONE
+
+        val programStagesHaveCoordinates by lazy {
+            searchRepository.programStagesHaveCoordinates(program.uid())
+        }
+
+        val programAttributesHaveCoordinates by lazy {
+            searchRepository.programAttributesHaveCoordinates(program.uid())
+        }
+
+        val teTypeHasCoordinates by lazy {
+            searchRepository.trackedEntityType?.let { teType ->
+                teType.featureType() != null && teType.featureType() != FeatureType.NONE
+            } ?: false
+        }
+
+        val teAttributesHaveCoordinates by lazy {
+            searchRepository.teTypeAttributesHaveCoordinates(program.trackedEntityType()?.uid())
+        }
+
+        val eventsHaveCoordinates by lazy {
+            searchRepository.eventsHaveCoordinates(program.uid())
+        }
+
+        return programHasCoordinates ||
+            programStagesHaveCoordinates ||
+            programAttributesHaveCoordinates ||
+            teTypeHasCoordinates ||
+            teAttributesHaveCoordinates ||
+            eventsHaveCoordinates
     }
 }