--- a/app/src/main/java/org/dhis2/usescases/searchTrackEntity/SearchRepository.java
+++ b/app/src/main/java/org/dhis2/usescases/searchTrackEntity/SearchRepository.java
@@ -13,6 +13,7 @@ import org.dhis2.data.search.SearchParametersModel;
 import org.hisp.dhis.android.core.arch.call.D2Progress;
 import org.hisp.dhis.android.core.organisationunit.OrganisationUnit;
 import org.hisp.dhis.android.core.program.Program;
+import org.hisp.dhis.android.core.settings.AnalyticsDhisVisualizationsGroup;
 import org.hisp.dhis.android.core.trackedentity.TrackedEntityType;
 import org.jetbrains.annotations.NotNull;
 
@@ -58,11 +59,12 @@ public interface SearchRepository {
     TeiDownloadResult download(String teiUid, @Nullable String enrollmentUid, String reason);
 
     void setCurrentProgram(@Nullable String currentProgram);
-    boolean programHasAnalytics();
     boolean programStagesHaveCoordinates(String programUid);
     boolean teTypeAttributesHaveCoordinates(String typeId);
     boolean programAttributesHaveCoordinates(String programUid);
     boolean eventsHaveCoordinates(String programUid);
+
+    List<AnalyticsDhisVisualizationsGroup> getProgramVisualizationGroups(String programUid);
 
     @Nullable Program getProgram(@Nullable String programUid);
     @Nullable String currentProgram();