## Description
There are two expensive evaluations performed before displaying the Program home screen (the TEI list). They are the evaluation of `displayMap` and `displayAnalytics`, which determined if these tabs must be shown in the bottom menu. This PR improves the performance of both variables.

[ANDROAPP-5660](https://dhis2.atlassian.net/browse/ANDROAPP-5660)

## Solution description
Improvements for displayMap:
- Early return when the conditions are enough to determine if the result is true or false.
- Single DB query instead of traversing all the events in the program.

Some measurements:
- Physical device, Android 8, program without coordinates with ~2000 teis: it improves from ~8s to ~0.070s (no early return, improvement thanks to single DB query).
- Physical device, Android 8, program with coordinates with 50 teis: it improves from ~0.300s to ~0.005s (thanks to early return to true).

Improvements for displayAnalytics:
- Do not evaluate the visualizations, just check if there are visualizations defined for this program. The evaluation was needed in previous versions because failed visualizations were skipped from the list and you could end up with an empty list. Now, they are not skipped anymore but displayed with an error message.

Some measurements:
- Physical device, Android 8: the improvement goes from the time it takes the evaluation (wide range, from 0.3s to several minutes) to just ~0.005s (it only checks the configuration).

## Covered unit test cases
The unit test proves that there is an early return.
## Where did you test this issue?
- [X] Smartphone Emulator
- [ ] Tablet Emulator
- [X] Smartphone
- [ ] Tablet
## Which Android versions did you test this issue?
- [ ] 4.4
- [ ] 5.X - 6.X
- [ ] 7.X
- [X] 8.X
- [ ] 9.X - 10.X
- [X] 11.X - 13.X
- [ ] Other
## Checklist
- [ ] My code follows the style guidelines of this project
- [ ] I have performed a self-review of my own code


[ANDROAPP-5660]: https://dhis2.atlassian.net/browse/ANDROAPP-5660?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ